requirements.txt
flask==3.0.3
flask-cors==4.0.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
python-dotenv==1.0.1
httpx==0.27.2
fastapi==0.115.0
uvicorn==0.30.6
gunicorn==22.0.0